requirements.txt
Django==3.2.16
django-after-response==0.2.2
git+https://github.com/ministryofjustice/opg-response.git#egg=django-incident-response
django-health-check==3.16.5
django-decorator-include
pytest-mock
pytest-django==4.5.2
python-dotenv==0.20.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
social-auth-app-django
statuspageio
pypd
